refactor(#5): DTO 생성 및 패키지 위치 변경
