+++ src/main/java/com/modive/llm/dto/request/DrivingSummaryRequest.java
@@ -1,11 +1,9 @@
-package com.modive.llm.dto;
+package com.modive.llm.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @Data
 @Builder
