+++ src/main/java/com/modive/llm/dto/request/FeedbackRequest.java
@@ -1,4 +1,4 @@
-package com.modive.llm.dto;
+package com.modive.llm.dto.request;
 
 import lombok.Builder;
 import lombok.Getter;
