+++ src/main/java/com/modive/llm/dto/response/DrivingSummaryResponse.java
@@ -1,4 +1,4 @@
-package com.modive.llm.dto;
+package com.modive.llm.dto.response;
 
 
 import lombok.AllArgsConstructor;
